+++ src/error.rs
@@ -1,7 +1,9 @@
 use solana_client::client_error::ClientError;
 use solana_sdk::pubkey::ParsePubkeyError;
 use thiserror::Error;
+use tokio::sync::oneshot::error::RecvError;
 use tokio::task::JoinError;
+use tokio::time::error::Elapsed;
 use tokio_tungstenite::tungstenite::Error as WsError;
 
 #[derive(Debug, Error)]
@@ -20,6 +22,12 @@ pub enum Error {
 
   #[error("WebSocket error")]
   WebSocketError(#[from] WsError),
+
+  #[error("AsyncWrap error")]
+  AsyncWrapError(#[from] RecvError),
+
+  #[error("AsyncTimeout error")]
+  AsyncTimeoutError(#[from] Elapsed),
 
   #[error("Notification for an unknown subscription")]
   UnknownSubscription,